requirements.txt
Flask
Flask-Cors
supabase
python-dotenv
requests
bs4
lxml